api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
motor[srv]>=3.4
pydantic>=2.7
python-dotenv>=1.0
